Move init, build free functions into 'builder.rs'
